fix bug due to numpy 1.19 changes + other small
* numpy 1.19 introduced the need to specify dtype=object for ragged arrays constructed from lists of lists. This broke part of the code (probably more bugs will come up). Edits in classes.LEEDPattern solve this in one instance (specify datatype explicitly).
* One new pattern definition file [KTO(001)] and exported pattern file

+++ guilib/leedsim/classes.py
@@ -149,7 +149,7 @@ class LEEDPattern():
         # beamsRefs contains at index [i, j] a list of the indices of beams of
         # domain domains[i] equivalent to beam j (including j)
         beamsRefs = []
-        # for dom range(nDoms):  # THIS IS PROBABLY WRONG FOR A LIST OF DOMAINS
+        # for dom in range(nDoms):  # THIS IS PROBABLY WRONG FOR A LIST OF DOMAINS
         for dom in domains: 
             beamsRefs.append([crossRef + dom*self.nBeams
                               for crossRef in self.firstDomRefs])
@@ -180,7 +180,7 @@ class LEEDPattern():
         #        at the same positional index.
         #        !!!THIS IS A PROBLEM FOR NON-NORMAL INCIDENCE!!!
         #        --> get this list (the same for all domains) ready first
-        allRefs = [np.concatenate(beamRefs[:, beam])
+        allRefs = [np.concatenate(beamRefs[:, beam]).astype(int)
                    for beam in range(self.nBeams)]
         refNames = [set(self.names.ravel()[beamRef]) for beamRef in allRefs]
